requirements.txt
openai
openai_secret_manager
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client